Update hash.go
Hashing should not rely on ProfoBuf encoding, as it could be
subject to change in the future (and its implementation is not
designed with security in mind).

The new hash() function accepts only a specific set of
specific-size integers, strings, and N-dimensional slices of those
types (with N >= 1). It takes care of endianess using the binary
built-in packageUpdate hash.go

Hashing should not rely on ProfoBuf encoding, as it could be
subject to change in the future (and its implementation is not
designed with security in mind).

The new hash() function accepts only a specific set of
specific-size integers, strings, and N-dimensional slices of those
types (with N >= 1). It takes care of endianess using the binary
built-in package..