Update collection.go, node.go
First scaffolding needed to implement temporary nodes
and transactions. Pointers to temporary nodes are now
stored (this will happen when verifying something
throughout a transaction of while applying updates).

Moreover, we will implement lazy updating to implement
batch update processing. This introduces
inconsistencies that can be signaled using a boolean
flag on the nodes. Inconsistencies are fixed using
the fix method on a collection.