Update collection.go, add proof.go
Adds basic implementation of proof, which is made
by a key, a root label and a set of steps. Each step
is made by two dumps, which in turn represents the
content of a node out of the tree context.

Adds Get().Proof() to collections, which returns a
proof that can be both of existence or non-existence.