Update node.go, transaction_tests.go
Extends transaction tests to account for Set and
Remove manipulators during transactions. New tests
showed that the backup() operation on node is not
safe, as it copies the values slice instead of its
content (therefore the backup values only act as a
reference to the values that get manipulated). The
issue is fixed by using the copy() function instead
of the assignment operator.