Add verifiers.go, update proof.go
Adds Verify() method to verify and extend the
collection tree given new proofs. The main difference
with the experimental branch is that no collection
is carried out at the end of the verification when
transaction is false. This means that collection only
occurs after a manipulator is called, when transaction
is false. This allows the user to verify a proof,
then call a manipulator without having to run a
whole transaction.

Fixes a bug in proof.go: method to() for a dump did
not set known to true on the target node.