Add update.go
Implements Update, ReadOnly and ReadWrite interfaces.
Update paradigm will be in three steps, namely
Records(), which is a getter for the proofs involved
in the update, Check() which will explicitly have
read-only access to the collection and serves the
purpose to verify if the update can be applied, and
Apply(). Both Check() and Apply() will cause a panic
if they try to access the collection over keys that
where not previously declared.

Adds proxy struct, which can be generated by a
collection and serves the sole purpose to implement
the ReadOnly and ReadWrite interfaces and filter
requests to records that are not whitelisted.